fix gateway exec tty cleanup on context.Canceled
This fixes an issue where the tty message handling loop will go into a
tight loop and never exit upon context.Canceled.  There is select
statement in `(*procMessageForwarder).Recv` that returns nil on
ctx.Done, but the control loop in `(*container).Start` did not exit on
this condition.  I think the intent was to flush out any inflight
messages on cancel, but this is already done in
`(*procMessageForwarder) Close`.

Signed-off-by: coryb <cbennett@netflix.com>